3_재택근무_관리시스템/requirements.txt
# Remote Work Management System - Dependencies
# Python 3.10+

# Core Framework
streamlit==1.32.0

# Database
# (SQLite3 is built-in)

# Security
argon2-cffi==23.1.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
python-dateutil==2.8.2

# Utilities
Pillow==10.2.0
